backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
ollama==0.1.7
pinecone-client==3.0.0
psycopg2-binary==2.9.9
pydantic>=2.5.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
python-jose[cryptography]==3.3.0
authlib==1.2.1
faster-whisper==1.0.0
google-generativeai==0.8.3
zhipuai==2.0.1
Pillow==10.0.0